Updated examples for digit and started on cpy example

+++ exampleprograms/patchedDigit2.c
@@ -2,10 +2,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-//Program only takes in positive numbers between 0-9
 int main(int argc, char *argv[]) {
 
-    int input = atoi(argv[1]);
+    long int input = atoi(argv[1]);
     long int overflow = 127; //-128 to 128
 
     long int test1 = overflow + input;
